examples/miniapps/sanic/requirements.txt
dependency-injector
sanic
sanic-testing
aiohttp
pyyaml
pytest-cov
